refactor(feature): clean up conditional code in updateState
The code in feature::updateState leads with a negative condition in the
initial branch, negating it in subsequent branches. It's easier for a
reader to start with a positive condition and then negate that. Also,
the code was re-asserting the negation of preceding conditions in
subsequent `if` statements, instead of using `else` or an early `return`
to let the flow happen naturally.

This change flips the conditions so that it leads with a positive
assertion and then uses `else` to reduce the logic in subsequent ones.